restructure model.user
