+++ insta-comments-scraper/comments-scraper.go
@@ -88,8 +88,8 @@ func (p *PostCommentScraper) Run() {
 	}
 }
 
-func (p *PostCommentScraper) scrapeComments(shortCode string) (*models.InstaPostComments, error) {
-	var postsComments *models.InstaPostComments
+func (p *PostCommentScraper) scrapeComments(shortCode string) (*InstaPostComments, error) {
+	var postsComments *InstaPostComments
 	err := p.httpClient.WithRetries(3, func() error {
 		time.Sleep(1400 * time.Millisecond)
 		instaPostComments, err := p.scrapePostComments(shortCode)
@@ -104,7 +104,7 @@ func (p *PostCommentScraper) scrapeComments(shortCode string) (*models.InstaPost
 	return postsComments, err
 }
 
-func (p *PostCommentScraper) sendComments(postsComments *models.InstaPostComments, postId models.InstagramPost) error {
+func (p *PostCommentScraper) sendComments(postsComments *InstaPostComments, postId models.InstagramPost) error {
 	fmt.Println("sendComments: ", len(postsComments.Data.ShortcodeMedia.EdgeMediaToParentComment.Edges))
 	messages := make([]kafka.Message, 0, len(postsComments.Data.ShortcodeMedia.EdgeMediaToParentComment.Edges))
 	for _, element := range postsComments.Data.ShortcodeMedia.EdgeMediaToParentComment.Edges {
@@ -131,8 +131,8 @@ func (p *PostCommentScraper) sendComments(postsComments *models.InstaPostComment
 	return p.commentsInfoQWriter.WriteMessages(context.Background(), messages...)
 }
 
-func (p *PostCommentScraper) scrapePostComments(shortCode string) (models.InstaPostComments, error) {
-	var instaPostComment models.InstaPostComments
+func (p *PostCommentScraper) scrapePostComments(shortCode string) (InstaPostComments, error) {
+	var instaPostComment InstaPostComments
 	type Variables struct {
 		Shortcode           string `json:"shortcode"`
 		ChildCommentCount   int    `json:"child_comment_count"`
