+++ insta-comments-scraper/insta-model.go
@@ -1,0 +1,144 @@
+package insta_comments_scraper
+
+type InstaPostComments struct {
+	Data struct {
+		ShortcodeMedia struct {
+			Typename   string `json:"__typename"`
+			ID         string `json:"id"`
+			Shortcode  string `json:"shortcode"`
+			Dimensions struct {
+				Height int `json:"height"`
+				Width  int `json:"width"`
+			} `json:"dimensions"`
+			GatingInfo           interface{} `json:"gating_info"`
+			FactCheckInformation interface{} `json:"fact_check_information"`
+			MediaPreview         string      `json:"media_preview"`
+			DisplayURL           string      `json:"display_url"`
+			DisplayResources     []struct {
+				Src          string `json:"src"`
+				ConfigWidth  int    `json:"config_width"`
+				ConfigHeight int    `json:"config_height"`
+			} `json:"display_resources"`
+			AccessibilityCaption  string `json:"accessibility_caption"`
+			IsVideo               bool   `json:"is_video"`
+			TrackingToken         string `json:"tracking_token"`
+			EdgeMediaToTaggedUser struct {
+				Edges []struct {
+					Node struct {
+						User struct {
+							FullName      string `json:"full_name"`
+							ID            string `json:"id"`
+							IsVerified    bool   `json:"is_verified"`
+							ProfilePicURL string `json:"profile_pic_url"`
+							Username      string `json:"username"`
+						} `json:"user"`
+						X float64 `json:"x"`
+						Y float64 `json:"y"`
+					} `json:"node"`
+				} `json:"edges"`
+			} `json:"edge_media_to_tagged_user"`
+			EdgeMediaToCaption struct {
+				Edges []struct {
+					Node struct {
+						Text string `json:"text"`
+					} `json:"node"`
+				} `json:"edges"`
+			} `json:"edge_media_to_caption"`
+			CaptionIsEdited          bool `json:"caption_is_edited"`
+			HasRankedComments        bool `json:"has_ranked_comments"`
+			EdgeMediaToParentComment struct {
+				Count    int `json:"count"`
+				PageInfo struct {
+					HasNextPage bool        `json:"has_next_page"`
+					EndCursor   interface{} `json:"end_cursor"`
+				} `json:"page_info"`
+				Edges []struct {
+					Node struct {
+						ID              string `json:"id"`
+						Text            string `json:"text"`
+						CreatedAt       int    `json:"created_at"`
+						DidReportAsSpam bool   `json:"did_report_as_spam"`
+						Owner           struct {
+							ID            string `json:"id"`
+							IsVerified    bool   `json:"is_verified"`
+							ProfilePicURL string `json:"profile_pic_url"`
+							Username      string `json:"username"`
+						} `json:"owner"`
+						ViewerHasLiked bool `json:"viewer_has_liked"`
+						EdgeLikedBy    struct {
+							Count int `json:"count"`
+						} `json:"edge_liked_by"`
+						EdgeThreadedComments struct {
+							Count    int `json:"count"`
+							PageInfo struct {
+								HasNextPage bool        `json:"has_next_page"`
+								EndCursor   interface{} `json:"end_cursor"`
+							} `json:"page_info"`
+							Edges []interface{} `json:"edges"`
+						} `json:"edge_threaded_comments"`
+					} `json:"node"`
+				} `json:"edges"`
+			} `json:"edge_media_to_parent_comment"`
+			EdgeMediaPreviewComment struct {
+				Count int `json:"count"`
+				Edges []struct {
+					Node struct {
+						ID              string `json:"id"`
+						Text            string `json:"text"`
+						CreatedAt       int    `json:"created_at"`
+						DidReportAsSpam bool   `json:"did_report_as_spam"`
+						Owner           struct {
+							ID            string `json:"id"`
+							IsVerified    bool   `json:"is_verified"`
+							ProfilePicURL string `json:"profile_pic_url"`
+							Username      string `json:"username"`
+						} `json:"owner"`
+						ViewerHasLiked bool `json:"viewer_has_liked"`
+						EdgeLikedBy    struct {
+							Count int `json:"count"`
+						} `json:"edge_liked_by"`
+					} `json:"node"`
+				} `json:"edges"`
+			} `json:"edge_media_preview_comment"`
+			CommentsDisabled     bool `json:"comments_disabled"`
+			TakenAtTimestamp     int  `json:"taken_at_timestamp"`
+			EdgeMediaPreviewLike struct {
+				Count int           `json:"count"`
+				Edges []interface{} `json:"edges"`
+			} `json:"edge_media_preview_like"`
+			EdgeMediaToSponsorUser struct {
+				Edges []interface{} `json:"edges"`
+			} `json:"edge_media_to_sponsor_user"`
+			Location struct {
+				ID            string `json:"id"`
+				HasPublicPage bool   `json:"has_public_page"`
+				Name          string `json:"name"`
+				Slug          string `json:"slug"`
+				AddressJSON   string `json:"address_json"`
+			} `json:"location"`
+			ViewerHasLiked             bool `json:"viewer_has_liked"`
+			ViewerHasSaved             bool `json:"viewer_has_saved"`
+			ViewerHasSavedToCollection bool `json:"viewer_has_saved_to_collection"`
+			ViewerInPhotoOfYou         bool `json:"viewer_in_photo_of_you"`
+			ViewerCanReshare           bool `json:"viewer_can_reshare"`
+			Owner                      struct {
+				ID                string `json:"id"`
+				IsVerified        bool   `json:"is_verified"`
+				ProfilePicURL     string `json:"profile_pic_url"`
+				Username          string `json:"username"`
+				BlockedByViewer   bool   `json:"blocked_by_viewer"`
+				FollowedByViewer  bool   `json:"followed_by_viewer"`
+				FullName          string `json:"full_name"`
+				HasBlockedViewer  bool   `json:"has_blocked_viewer"`
+				IsPrivate         bool   `json:"is_private"`
+				IsUnpublished     bool   `json:"is_unpublished"`
+				RequestedByViewer bool   `json:"requested_by_viewer"`
+			} `json:"owner"`
+			IsAd                       bool `json:"is_ad"`
+			EdgeWebMediaToRelatedMedia struct {
+				Edges []interface{} `json:"edges"`
+			} `json:"edge_web_media_to_related_media"`
+		} `json:"shortcode_media"`
+	} `json:"data"`
+	Status string `json:"status"`
+}
